+++ ui/pages/Search/components/VariantSearchResults.jsx
